Merging Request

+++ src/main/java/com/aktic/linkWhiz_backend/service/user/UserService.java
@@ -4,6 +4,7 @@ import com.aktic.linkWhiz_backend.model.entity.User;
 import com.aktic.linkWhiz_backend.model.request.ChangePasswordRequest;
 import com.aktic.linkWhiz_backend.model.response.UserInfo;
 import com.aktic.linkWhiz_backend.repository.UserRepository;
+import com.aktic.linkWhiz_backend.service.auth.AuthService;
 import com.aktic.linkWhiz_backend.service.fileStorage.FileStorageService;
 import com.aktic.linkWhiz_backend.util.ApiResponse;
 import lombok.RequiredArgsConstructor;
@@ -13,16 +14,12 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.nio.file.Files;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -32,20 +29,11 @@ public class UserService {
     private final UserRepository userRepository;
     private final FileStorageService fileStorageService;
     private final PasswordEncoder passwordEncoder;
+    private final AuthService authService;
 
     public ResponseEntity<ApiResponse<UserInfo>> updateProfile(String firstName, String lastName, MultipartFile image) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
-
         try {
-            Optional<User> optionalUser = userRepository.findByEmail(userDetails.getUsername());
-
-            if (optionalUser.isEmpty()) {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                        .body(new ApiResponse<>(false, "User not found", null));
-            }
-
-            User user = optionalUser.get();
+            User user = authService.getCurrentUser();
 
             if (firstName != null) {
                 user.setFirstName(firstName);
@@ -54,7 +42,7 @@ public class UserService {
                 user.setLastName(lastName);
             }
             if (image != null) {
-                fileStorageService.delete(user.getImage());
+                if (user.getImage() != null) fileStorageService.delete(user.getImage());
                 user.setImage(fileStorageService.save(image));
             }
 
@@ -71,16 +59,7 @@ public class UserService {
 
     public ResponseEntity<Resource> getProfileImage() {
         try {
-            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-            UserDetails userDetails = (UserDetails) authentication.getPrincipal();
-
-            Optional<User> optionalUser = userRepository.findByEmail(userDetails.getUsername());
-
-            if (optionalUser.isEmpty()) {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-            }
-
-            User user = optionalUser.get();
+            User user = authService.getCurrentUser();
 
             Resource resource = fileStorageService.load(user.getImage());
             if (resource == null || !resource.exists()) {
@@ -108,8 +87,7 @@ public class UserService {
 
     public ResponseEntity<ApiResponse<String>> updatePassword(ChangePasswordRequest changePasswordRequest) {
         try {
-            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-            User user = (User) authentication.getPrincipal();
+            User user = authService.getCurrentUser();
 
             if (changePasswordRequest.getOldPassword().equals(changePasswordRequest.getNewPassword())) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST)
